requirements.txt
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
Deprecated==1.2.11
future==0.18.2
graphtools==1.5.2
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
littleutils==0.2.2
llvmlite==0.33.0
magic-impute==2.0.4
matplotlib==3.3.4
networkx==2.5
numba==0.50.1
numpy==1.19.5
outdated==0.2.0
packaging==20.9
pandas==1.1.5
pandas-flavor==0.2.0
patsy==0.5.1
Pillow==8.1.0
pingouin==0.3.10
pkg-resources==0.0.0
PyGSP==0.5.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.5.4
scprep==1.0.13
seaborn==0.11.1
six==1.15.0
statsmodels==0.12.2
tabulate==0.8.7
tasklogger==1.0.0
threadpoolctl==2.1.0
tqdm==4.56.2
urllib3==1.26.3
wrapt==1.12.1
xarray==0.16.2
